requirements.txt
langchain-core
langchain-openai
streamlit
openai
tiktoken
python-dotenv
faiss-cpu